Add files via upload
PROJECT OVERVIEW
Our project is the implementation of the classic Tic Tac Toe game using HTML, CSS, and JavaScript. The game involves two players who take turns marking squares on a 3x3 grid. The objective of the game is to be the first player to get three of their marks in a row, either horizontally, vertically, or diagonally.

Technologies Used:

We have used three web development technologies for this project:

HTML: It provides the structure and content of the web page.
CSS: It is used for styling and visual presentation of the game interface.
JavaScript: It provides the game functionality and interactivity, such as capturing player moves, checking for wins, and updating the game board.

Features and Functionalities:

Our Tic Tac Toe game implementation includes the following features and functionalities:

1. Two players can play the game locally on the same device.
2. Players can take turns marking squares on a 3x3 grid by clicking on them.
3. The game logic checks for winning conditions after each move and displays a winning message when a player wins.
4. The game also handles the case of a tie or draw, where no player wins.
5. The game includes a reset button that allows players to start a new game              
6. The game interface is designed to be visually appealing, with custom CSS styling, animations, and responsive design for different screen sizes.
7. Overall, our Tic Tac Toe game implementation provides a fun and engaging    experience for players, while also demonstrating our proficiency in HTML, CSS, and JavaScript web development.